Extract color picker methods into separate utilities
