+++ apps/client/src/components/color-picker.tsx
@@ -8,82 +8,12 @@ import {
 } from "@/components/ui/popover";
 import { Slider } from "@/components/ui/slider";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { type HSL, hslToRgb, type RGB, rgbToHex, rgbToHsl } from "@/lib/color";
+import { clamp } from "@/lib/math";
 import { cn } from "@/lib/utils";
 
-type RGB = { r: number; g: number; b: number };
-type HSL = { h: number; s: number; l: number };
-
-function clamp(v: number, a = 0, b = 255) {
-	return Math.min(Math.max(v, a), b);
-}
-
-function rgbToHex({ r, g, b }: RGB) {
-	const toHex = (n: number) => n.toString(16).padStart(2, "0");
-	return `#${toHex(r)}${toHex(g)}${toHex(b)}`;
-}
-
-function rgbToHsl({ r, g, b }: RGB): HSL {
-	r /= 255;
-	g /= 255;
-	b /= 255;
-	const max = Math.max(r, g, b);
-	const min = Math.min(r, g, b);
-	let h = 0;
-	let s = 0;
-	const l = (max + min) / 2;
-
-	if (max !== min) {
-		const d = max - min;
-		s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-		switch (max) {
-			case r:
-				h = (g - b) / d + (g < b ? 6 : 0);
-				break;
-			case g:
-				h = (b - r) / d + 2;
-				break;
-			case b:
-				h = (r - g) / d + 4;
-				break;
-		}
-		h /= 6;
-	}
-	return {
-		h: Math.round(h * 360),
-		s: Math.round(s * 100),
-		l: Math.round(l * 100),
-	};
-}
-
-function hslToRgb({ h, s, l }: HSL): RGB {
-	h = h / 360;
-	s = s / 100;
-	l = l / 100;
-
-	if (s === 0) {
-		const v = Math.round(l * 255);
-		return { r: v, g: v, b: v };
-	}
-
-	const hue2rgb = (p: number, q: number, t: number) => {
-		if (t < 0) t += 1;
-		if (t > 1) t -= 1;
-		if (t < 1 / 6) return p + (q - p) * 6 * t;
-		if (t < 1 / 2) return q;
-		if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
-		return p;
-	};
-
-	const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-	const p = 2 * l - q;
-	const r = Math.round(hue2rgb(p, q, h + 1 / 3) * 255);
-	const g = Math.round(hue2rgb(p, q, h) * 255);
-	const b = Math.round(hue2rgb(p, q, h - 1 / 3) * 255);
-	return { r, g, b };
-}
-
 export function ColorPicker({
-	initial = { r: 110, g: 170, b: 230 },
+	initial = { r: 255, g: 230, b: 0 },
 	className,
 	onChange,
 	optional = false,
