+++ apps/client/src/lib/math.ts
@@ -1,0 +1,6 @@
+/**
+ * Clamp a value between a minimum and maximum
+ */
+export function clamp(value: number, min = 0, max = 255): number {
+	return Math.min(Math.max(value, min), max);
+}
